X64 asm

+++ lib/std/sample.go
@@ -1,3 +1,3 @@
-func value() int {
+func value() int64 {
     return 30;
 }
